// src/api/generated/model/mgmtRolloutRestRequestBodyPost.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtRolloutCondition } from './mgmtRolloutCondition';
import type { MgmtRolloutSuccessAction } from './mgmtRolloutSuccessAction';
import type { MgmtRolloutErrorAction } from './mgmtRolloutErrorAction';
import type { MgmtDynamicRolloutGroupTemplate } from './mgmtDynamicRolloutGroupTemplate';
import type { MgmtRolloutRestRequestBodyPostType } from './mgmtRolloutRestRequestBodyPostType';
import type { MgmtRolloutGroup } from './mgmtRolloutGroup';
import type { Links } from './links';

export interface MgmtRolloutRestRequestBodyPost {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  successCondition?: MgmtRolloutCondition;
  successAction?: MgmtRolloutSuccessAction;
  errorCondition?: MgmtRolloutCondition;
  errorAction?: MgmtRolloutErrorAction;
  /** Target filter query language expression */
  targetFilterQuery?: string;
  /** The ID of distribution set of this rollout */
  distributionSetId?: number;
  /** The amount of groups the rollout should split targets into */
  amountGroups?: number;
  /** Force time in milliseconds */
  forcetime?: number;
  /** Start at timestamp of Rollout */
  startAt?: number;
  /** Weight of the resulting Actions */
  weight?: number;
  dynamic?: boolean;
  dynamicGroupTemplate?: MgmtDynamicRolloutGroupTemplate;
  /** (Available with user consent flow active) If the confirmation is required for this rollout. Value will be used
if confirmation options are missing in the rollout group definitions. Confirmation is required per default */
  confirmationRequired?: boolean;
  /** The type of this rollout */
  type?: MgmtRolloutRestRequestBodyPostType;
  /** The list of group definitions */
  groups?: MgmtRolloutGroup[];
  _links?: Links;
}